Adds Randomizer assembly.
This acts a separate library which does *not* have a dependency on Unity. It does appear to be using .NET Framework, when I would like it to use .NET Standard, but I'm not sure how to change that right now.

For now, this just exposes a single function and calls it from the default Unity assembly.

https://docs.unity3d.com/Manual/ScriptCompilationAssemblyDefinitionFiles.html