Adds simple randomization and spawns items at the appropriate checks.
This adds a representation of `Check` objects and ties them to the existing `Item` objects. Each `Check` has a location and a named enum value which gets randomly mapped to a prefab for the associated `Item`, which gets instantiated at the `Check` location.

For now, the randomization algorithm is as simple as possible. It simply picks a purely random item for every check, with no logic or validation. Items can and will repeat across checks and some may be missing altogether.