Defines the logic graph and parses it from a YAML file.
The YAML file describes the game logic and how each check relates to other checks. This is conceptualized as a "lock and key" system crossed with a graph. Each check is a lock with a set of keys required to open it. However, there may be multiple ways to get to a given check and not all are accessible from game start. We express this with a graph of these checks, and each check is accessible from one or more other checks plus one of several possible sets of keys. It may be possible to use the bow to get to a check from one side, but it may also be possible to get to the same check from another side with a bomb and a sword. This system allows for expressing this kind of logic.

We start by processing all this into an immutable graph of logical nodes and edges with a named set of keys and nodes. The algorithmic challenge from here is to find a random placement of keys which is still logically solvable.

I left most of this in the Unity assembly just so I could leverage it to read the text file, which is otherwise a bit complicated to get a stable reference to. The algorithm should be independent of the logic graph, but the graph should also be separate from Unity so it can be used in other contexts (ie. WebAssembly). I'll deal with that part later however.

Ironically, the trickiest part of this is loading external packages. This requires two third-party packages, a YAML parser and an immutable collections library. Both of which are available on NuGet and compatible with Unity. However, apparently you just can't install NuGet packages into a Unity project. Since Unity owns the `.csproj` and `.sln` files, installed packages do not persist and Unity appears to have no direct support for this. The best workaround I could fine is to manually download the relevant package from NuGet and put its DLL into the `Assets\Packages\` directory where Unity will load it automatically. This is very suboptimal for a number of reasons, most of all that I need to be a human package manager and manually download all transitive dependencies and resolve them to compatible versions. D: