Initial randomization algorithm implementation.
This implements the general algorithm I initially had in mind. Documentation is mostly inline in the `Randomizer` class. This is terribly sub-optimal as it repeats a lot of work it doesn't need to. However, this is good enough to generate a few seeds which seem to solvable.

Beyond the randomization algorithm, I added a Bluepee item to act as junk and moved the `Item` and `Check` classes to the Unity project, as the randomizer should not have any checks or items hard-coded.