Adds basic inventory system.
This leverages ScriptableObjects to represent different kinds of items in the player's inventory. Currently all the existing items of sword, bow, boms, and rupees are represented and held in the player's inventory. Each item in the world has a reference to the inventory item that will be given to the player when picked up. This keeps dependencies convenient because each item has a dependency on the inventory and knows how to add itself to an inventory, but the inventory doesn't have knowledge abou specific items or how to get them. It only holds a reference to each kind of inventory item.

Each item has some basic metadata associated with it, though none of it is really used right now.