Changes `ImmutableHashSet<T>` usages to `ImmutableSortedSet<T>`.
Apparently, `ImmutableHashSet<T>` ordering is non-deterministic. Constructing two identical sets with inputs in the same order does *not* guarantee that iterating over the output will be in a consistent order. For the randomizer, we do expect determinism for a given seed, so while they are unordered sets, being non-deterministic is a big issue resulting in different outputs for the same seed. https://twitter.com/develwoutacause/status/1395160193999069189

Fixed this by changing all `ImmutableHashSet<T>` references to `ImmutableSortedSet<T>`. We don't actually need them to be sorted, just deterministic when enumerated, but this seems to be the best way I can find to do that. Unfortunately this means a lot of the `Logic*` types need to implement `IComparable<T>` in order to be usable in an `ImmutableSortedSet<T>`. Fortunately they reduce down to string values on the nodes and keys, so it mostly just composes string comparisons.