Adds walking animations.
This uses the Unity animation controller to create a graph of animations controlled by triggers fired by the `PlayerMovementBehavior`. Basically pressing the move key for any direction always triggers the walking animation for that direction. Releasing triggers the "Stop" event which moves to the associated idle direction.

This looks fairly professional for a 2D game like this. Even moving diagonally works as I would generally expect it to.