Adds sword slashing animation.
This got fairly complicated, but all it really does is allow the user to press a button (currently `<`) to slash the sword in the direction they are currently facing. This stops the player and triggers the relevant animation before resuming movement. I ended up making a few utility functions to compose this "stop, do an action, resume" behavior which I imagine will be relatively common going forward.

The feeling of these controls is "good enough". It still doesn't feel particularly fluid for a few reasons:
1. Mashing the slash button queues one additional slash, rather than dropping excess inputs mid-animation.
1. The player sprite moves slightly during the slash pose, the sprites should probably be offset to keep his position relatively stationary.