Adds damage to the player's sword and allows it to destroy bushes.
This defines two core primitive prefabs, a `HitBox` and a `HurtBox`. A `HitBox` has an associated damage stat, while a `HurtBox` as an associated defense stat. Whenever these two collide, the `HitBox` damage is applied to the `HurtBox` defense, and any left-over damage is passed on to the target of the `HurtBox` and its associated `HealthBehavior`, which destroys the object if it reaches zero.

This can be leveraged in any prefab by simply making a child `HitBox` or `HurtBox` object, positioning the collider, and giving them damage / defense values. This should make it easy to construct various enemies and attacks without requiring custom code for each one (hooray!).

The actual implementation is simply a couple behaviors (`AttackBehavior` and `DefenseBehavior`) which hold an associated `Attack` / `Defense` stat. On any collision, these values are extracted and apply to the target object (usually the parent of the `HurtBox`).

I also modeled a separate `Resistence` ADT, which allows actors to define special resistences to particular attack types. For now, there is only a base attack and base defense value, but we could add a fire attack which applies double damage to ice-based enemies, but then ice attacks may do half damage or none at all. Different kinds of resistences can define this logic, but for now it's just a simple reduction of the input attack values.

The player doesn't yet have a `HurtBox`, but each sword direction does have a rough `HitBox`, so it is possible to cut grass with the sword now.