Scales sprites correctly.
Sprites were previously scaled as either 100 pixels / unit or 32 pixels / unit, which made them inconsistent and not fit very well in the Unity grid. This updates them to use 16 pixels / unit, which maps a single bush to a 1x1 square in Unity.

I also updated the bush prefab to shift the sprite and its colliders down and right by .5 units. This makes the origin the top left corner, so it aligns to the grid much more easily. Unfortunately, I can't seem to get this to apply for all objects very easily. I have to separate the sprite from the bush object to give it a distinct transform under the prefab in order to create this effect.