Tiles world.
This adds a tilemap and converts all the world sprites into tiles that can be easily. I say "easily", getting this working was a massive pain in the ass and the sprites aren't all a consistent size, so the tile palette is really awkward to work with.

I also adjusted the sprite resolution again. I changed it to 8 pixels / unit, as the smallest tiles were 8x8 pixels, and leave a blank space for the remaining pixels if placed on a 16x16 grid. This again means that I had to rescale everything and adjust hitboxes and whatnot.

The camera unfortunately displays vertical lines as weird artifacts despite my attempts at turning off all anti-aliasing, which is supposed to fix that. I suspect the camera is not aligned with the sprite resolution and causes it to draw "between" the sprites, creating this artifact.