Adds bombs.
This allows the player to press a button "." to place a bomb in front of them. The bomb will pulse for a few frames, and then explode, dealing some damage to any nearby objects.

The actual bomb logic is straightforward, it just waits a little while, then spawns and explosion and deletes itself. The explosion is even simpler, it just waits a little while and deletes itself. Both of these are done based on animation timing, and the explosion is only one with a hitbox. It also just deals base damage for now, but will deal explosive damage in the future.

The most complicated part is placing a bomb in front of the player. This requires a bit of math to find a bomb-sized space directly in front of the player. I couldn't find a good way to compute the bomb's collider size from its prefab, as it seems to always have size 0, likely because the bomb hasn't be instantiated yet. I just hard-coded the bomb's size for now, hopefully that won't change too frequently.

Also refactored some assets so player weapons are in their own folder, with sub folders for each weapon. Some of these assets leak into the general Player directory, but hopefully this provides a reasonable organization strategy.