Only allows placing bombs if the player has picked up the bombs item.
The actual requirement for the bombs item is trivial, however the tricky part is configuring placed bombs to use the same `Damage` object configured on the `BombInventoryItem`. We need to propagate this through to the explosion hit box, which is a few layers of indirection away. To make this more complicated, both bombs and explosions are spawned dynamically with `Instantiate()`, which does not accept user-defined parameters. Instead, the best way I could find was to write the inputs to a static field, and then read and validate that field in `Awake()`. This at least provides a useful error message if the developer gets it wrong.