Adds explosive damage and resistence and updates bombs to deal only explosive damage.
The change here is trivial, however it introduces a new resistence stat that nothing sets. This causes an NPE in `Defense.cs` with no simple answer. We can't easily default to a `SimpleResistence` of 0 because that requires dynamically constructing a `SimpleResistence` (or building one statically at startup), neither of which are well supported in Unity.

I think the proper solution here is similar to the `Damage` and `DamageConfig` distinction. We need one `ResistenceConfig` which extends `ScriptableObject` for use in Unity which simply creates `Resistence` objects respsonsible for the actual calculation. Then we can have a static `Resistence.Zero` property created on startup and referenced as a default value.

For now, I just made `Defense` check for `null` explicitly and fallback to a custom evaluation (the indentity function). It breaks abstraction somewhat, but it saves me the effort of making a non-`ScriptableObject` version of `Resistence` for right now.