Adds a rock which can only be destroyed via bombs.
The rock is a simple prefab with a hurtbox. Its most interesting trait is that it's defense is immune to base damage, meaning it is only vulnerable to explosive damage from bombs. This required a new `ImmuneResistence` which drops all received damage. Fortunately it has no parameters so it can easily be reused across multiple objects.