Adds the ability to shoot arrows.
This introduces a new `PlayerBowBehavior` which spawns an arrow in front of the player when they press the "Shoot Arrow" button (currently `.`, with bombs remapped to `/`). This spawns an arrow in front of the player at the right orientation.

The arrow itself simply has a hitbox to deal damage. It has an `ArrowBehavior` which controls moving the arrow forward each frame. Just like bombs and explosions, arrows also require some parameters on startup, however they actually require *two* parameters, damage and speed. I made a separate `ArrowBehavior.Params` class which is passed in via the static property side channel trick used by bombs and explosions. This allows the player's bow inventory item to influence the properties of fired arrows. A second bow upgrade, can easily increase arrow damage or speed.