Adds evasion stat.
This allows an object to include some evasion metadata. For now this just includes whether the object is "tall" or "short", which currently just means whether or not it would be hit by an arrow. For example, a bush is short enough to go under an arrow, while a rock/boulder is tall enough to still be hit by it.

I had to invert the dependency somewhat here. Previously, only the `DefenseBehavior` knew of the collision with an `AttackBehavior` object and received the damage. Then we had to add collision detection to the `AttackBehavior`, so it knew whether or not a hit had occurred and could trigger the `Hit` event. However, now it needed to know that the attack was evaded and `Hit` should *not* be triggered (so the arrow doesn't destroy itself). To do this, I removed the trigger from the `DefenseBehavior` and instead relied on `AttackBehavior` to depend on and call the `DefenseBehavior` appropriate. It has a `Receive()` method, which is basically what it before, however now it also has an `Evade()` method to check evasion first. This allows the `AttackBehavior` to check whether the defender evaded the hit, and only apply damage and register the hit if it failed to evade.

With this system, an arrow will fly harmlessly over a bush, but will collide with a rock, which is the main effect we really wanted to see.