Adds barriers for a couple checks.
This surrounds one check in bushes and another in rocks. The first requires either a sword or bombs to reach, while the second requires bombs. Also updated the logic file and check names to align.

With this, the randomizer makes a little more sense. Picking different seeds correctly randomizes items in a logically-compatible fashion, though most seeds seem to have all the progression items in sphere 0? Not sure if that's a statistical bug in the randomization algorithm or just being unlucky.