Adds a `Switch` prefab, which flips between a blue and red state after being struck.
The switch itself is relatively straightforward, just detecting a hit and changing its sprite. I couldn't find the sprites themselves on any website, instead I had to get some screenshots of Link to the Past from Tower or Hera and manually crop out the switch sprites.

This did require some refactoring for the `DefenseBehavior`. A switch doesn't have health and can't die, however `DefenseBehavior` expected to be protecting something with health. I couldn't find a *great* way of doing this, but I ultimately settled on an property to determine whether or not the `DefenseBehavior` has a target with health it is defending. This allows not setting a `target` property to be an assertion failure, but still support the case for hittable objects in the world that don't have "health" strictly speaking.

With this change, the `HealthBehavior` now has a `Hit` event, much like `AttackBehavior` which the `SwitchBehavior` is able to listen to and update itself accordingly.