Adds switched event to `SwitchBehavior`.
This shows an interface in the editor to call a funciton on another object with a boolean input. It sets the bridge state to be crossable, however for now it will always make the bridge crossable and will never hide it. This avoids the case where the player can drown by flipping the switch while on the bridge.