Add CS project generation fix.
I think Unity updated underneath me at some point and now editing the logic YAML opens in VSCode which causes a regeneration of the CS project, including `<ReferenceOutputAssembly>false</ReferenceOutputAssembly>` in the project reference for the Randomizer. This means nothing gets depended upon or can be imported.

This seems to just be a bug in Unity, but after updating everything and even downgrading and removing some relevant packages, I still see the issue. Best fix I found was a post-process step which forcibly removes this field. Not exactly elegant and shouldn't be needed to begin with, but it's the best I've got.