Adds title screen.
Ripped the sprite via a simple screenshot and edited the title very poorly. I think I was able to get the right font ("Triforce" from https://zeldauniverse.net/media/fonts/), though the shading and graidents make it difficult to get an exact match. Regardless, this is good enough for someone of my artistic ability.

When the player presses any key, it progresses to the overworld to start the game.