Adds instructions scene.
This simply includes some text about the controls and the ability to specify (or randomly generate) a seed value. This is loaded immediately after the title screen and has a "Start Game" button to continue to the Overworld.

This added TextMesh Pro which installed a bunch of junk in the project. I also added the Triforce font used in the title card (from https://zeldauniverse.net/media/fonts/) and used that for all player-visible text.

Currently, the seed value is ignored. Next step is to tie it into the actual game.