Uses the input seed to randomize the game.
This is a little awkward, but basically the Instructions scene is responsible for parsing the seed from user input and randomizing the game before changing to the Overworld scene. The one edge case here is for debug builds of the Overworld scene directly, which will lazily load a debug seed specified directly on the `Randomization` object.

This also shows an error message if it failed to parse an integer within the right bounds.