Always advances the title screen after 3 seconds.
This speeds up the auto-advance and removes the player's ability to press a key to move forward. Reasoning for this is because when the game ends, the player may be moving around and pressing buttons, which would immediately skip the title after it shows but before the player has a chance to see it. This is just the simplest solution to that problem.