Updates scenes to be compatible with a 960x600 resolution.
The WebGL target generates a canvas with that resolution and it seems to be non-trivial to change it (requires JavaScript on the page with a build process I don't want to set up or mess with).

Instead, I just "bought the gnome" and accepted that as the target resolution, updating the game to comply with that. The camera has the right background color for all three scenes, so a larger resolution should still look reasonable and be playable. Lower resolutions will start to cut things off.