Places all optional keys.
After placing the minimum required keys to reach all locations, some keys may be left over. For example, the sword is never required because bombs can also blast away grass. As a result, placing the bow and bombs first can make all location reachable on their own without ever having placed the sword. With this change, the sword will still be placed somewhere in the world, but won't be required to actually beat the seed.