Updates project settings to disable compression for WebGL builds.
With compression, the JavaScript files are compressed with Brotly and renamed to `*.js.br`. The `.br` extension confuses most servers and requires manual configuration. Since this is all being deployed on Netlify which compresses automatically, this isn't necessary in the first place, so the easiest option is to just disable it.